Typing Speed Game App 2015
Written in C in 2015. This program is a typing speed game. The user is
prompted to type a random word out of nine words in a sentence. The user
is timed. If the user fails to type the word correctly, the user is
prompted to try again. Once the user has typed all the words in the
sentence correctly the timer stops and the time the user took to
complete the task is displayed.